requirements.txt
selenium==4.16.0
pillow==10.1.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
